+++ src/letters/letters.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
-import { LettersController } from './letters.controller';
-import { LettersService } from './letters.service';
+import { LettersController } from './controller/letters.controller';
+import { LettersService } from './service/letters.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { LetterRepository } from './letters.repository';
-import { Letter } from './letters.entity';
+import { LetterRepository } from './repository/letters.repository';
+import { Letter } from './model/letters.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Letter])],
   controllers: [LettersController],
-  providers: [LetterRepository, LettersService],
+  providers: [LetterRepository, LettersService, LetterRepository],
 })
 export class LettersModule {}
