+++ src/letters/model/letter-scope.enum.ts
